sanity/requirements.txt
selenium
pyyaml
webdriver-manager
